refactor(enumerate): decompose the arp god-file by role (WS-D)
Split internal/discovery/enumerate/arp.go (750 LOC) into role-based files
in the same package — pure move: scanner type + config/accessors stay in
arp.go; subnet scanning + ARP-table reading/enrichment move out.

All files under the 600-line size-gate threshold. Declaration sets and
code-line multisets verified identical to the original.

Related to #187
